refactor: duplicated article card implementations
