+++ src/components/ArticleCard/index.tsx
@@ -1,0 +1,283 @@
+/** @jsx jsx */
+import { jsx } from 'theme-ui';
+import React from 'react';
+import styled from '@emotion/styled';
+import { css } from '@emotion/core';
+import { Link } from 'gatsby';
+
+import mediaqueries from '@styles/media';
+import { H2 } from '@components/Headings';
+import { IArticle } from '@types';
+import { GridLayoutContext } from '@sections/articles/ArticlesListContext';
+import Image, { ImagePlaceholder } from '@components/Image';
+
+interface IArticleProps {
+  article: IArticle;
+  narrow?: boolean;
+  isTiles?: boolean;
+}
+
+export default function ArticleCard({
+  article,
+  narrow,
+  isTiles,
+}: IArticleProps) {
+  const { gridLayout } = React.useContext(GridLayoutContext);
+  // eslint-disable-next-line no-param-reassign
+  isTiles = isTiles ?? gridLayout === 'tiles';
+  if (!article) {
+    return null;
+  }
+
+  const hasOverflow = narrow && article.title.length > 35;
+  const imageSource = narrow ? article.hero.narrow : article.hero.regular;
+  const hasHeroImage =
+    imageSource &&
+    Object.keys(imageSource).length !== 0 &&
+    imageSource.constructor === Object;
+
+  return (
+    <ArticleLink
+      to={article.slug}
+      data-a11y="false"
+      sx={{
+        '&:hover h2,&:focus h2': {
+          color: 'accent',
+        },
+        "&[data-a11y='true']:focus::after": {
+          borderColor: 'accent',
+        },
+      }}
+    >
+      <Article isTiles={isTiles}>
+        <ImageContainer narrow={narrow} isTiles={isTiles}>
+          {hasHeroImage ? <Image src={imageSource} /> : <ImagePlaceholder />}
+        </ImageContainer>
+        <div>
+          <Title
+            hasOverflow={hasOverflow}
+            isTiles={isTiles}
+            sx={{
+              fontFamily: 'serif',
+            }}
+          >
+            {article.secret ? `${article.title} 🚧` : article.title}
+          </Title>
+          <Excerpt
+            narrow={narrow}
+            hasOverflow={hasOverflow}
+            isTiles={isTiles}
+            sx={{
+              color: 'grey',
+            }}
+          >
+            {article.excerpt}
+          </Excerpt>
+          <MetaData
+            sx={{
+              color: 'grey',
+            }}
+          >
+            {article.date}
+            {/* · 阅读需要 {article.timeToRead} 分钟 */}
+          </MetaData>
+        </div>
+      </Article>
+    </ArticleLink>
+  );
+}
+
+const ImageContainer = styled.div<{ narrow: boolean; isTiles: boolean }>`
+  position: relative;
+  height: ${p => (p.isTiles ? '280px' : '220px')};
+  box-shadow: 0 30px 60px -10px rgba(0, 0, 0, ${p => (p.narrow ? 0.22 : 0.3)}),
+    0 18px 36px -18px rgba(0, 0, 0, ${p => (p.narrow ? 0.25 : 0.33)});
+  margin-bottom: ${p => (p.isTiles ? '30px' : 0)};
+  transition: transform 0.3s var(--ease-out-quad),
+    box-shadow 0.3s var(--ease-out-quad);
+
+  & > div {
+    height: 100%;
+  }
+
+  ${mediaqueries.tablet`
+    height: 200px;
+    margin-bottom: 35px;
+  `}
+
+  ${mediaqueries.phablet`
+    overflow: hidden;
+    margin-bottom: 0;
+    box-shadow: none;
+    border-top-right-radius: 5px;
+    border-top-left-radius: 5px;
+  `}
+`;
+
+const ArticleLink = styled(Link)`
+  position: relative;
+  display: block;
+  width: 100%;
+  height: 100%;
+  top: 0;
+  left: 0;
+  border-radius: 5px;
+  z-index: 1;
+  transition: transform 0.33s var(--ease-out-quart);
+  -webkit-tap-highlight-color: rgba(255, 255, 255, 0);
+
+  &:hover ${ImageContainer}, &:focus ${ImageContainer} {
+    transform: translateY(-1px);
+    box-shadow: 0 50px 80px -20px rgba(0, 0, 0, 0.27),
+      0 30px 50px -30px rgba(0, 0, 0, 0.3);
+  }
+
+  &[data-a11y='true']:focus::after {
+    content: '';
+    position: absolute;
+    left: -1.5%;
+    top: -2%;
+    width: 103%;
+    height: 104%;
+    background: rgba(255, 255, 255, 0.01);
+    border-radius: 5px;
+  }
+
+  ${mediaqueries.phablet`
+    &:hover ${ImageContainer} {
+      transform: none;
+      box-shadow: initial;
+    }
+
+    &:active {
+      transform: scale(0.97) translateY(3px);
+    }
+  `}
+`;
+
+const listItemRow = p => css`
+  display: grid;
+  grid-template-rows: 1fr;
+  grid-template-columns: 1fr 488px;
+  grid-column-gap: 96px;
+  grid-template-rows: 1;
+  align-items: center;
+  position: relative;
+  margin-bottom: 50px;
+
+  ${mediaqueries.desktop`
+    grid-column-gap: 24px;
+    grid-template-columns: 1fr 380px;
+  `}
+
+  ${mediaqueries.tablet`
+    grid-template-columns: 1fr;
+  `}
+
+  @media (max-width: 540px) {
+    background: ${p.theme.colors.card};
+  }
+
+  ${mediaqueries.phablet`
+    box-shadow: 0px 20px 40px rgba(0, 0, 0, 0.2);
+    border-bottom-right-radius: 5px;
+    border-bottom-left-radius: 5px;
+  `}
+`;
+
+const listItemTile = p => css`
+  position: relative;
+
+  ${mediaqueries.tablet`
+    margin-bottom: 60px;
+  `}
+
+  @media (max-width: 540px) {
+    background: ${p.theme.colors.card};
+  }
+
+  ${mediaqueries.phablet`
+    margin-bottom: 40px;
+    box-shadow: 0px 20px 40px rgba(0, 0, 0, 0.2);
+    border-bottom-right-radius: 5px;
+    border-bottom-left-radius: 5px;
+  `}
+`;
+
+const Article = styled.article<{ isTiles: boolean }>`
+  ${p => (p.isTiles ? listItemTile : listItemRow)}
+`;
+
+const limitToTwoLines = css`
+  text-overflow: ellipsis;
+  overflow-wrap: normal;
+  -webkit-line-clamp: 2;
+  -webkit-box-orient: vertical;
+  display: -webkit-box;
+  white-space: normal;
+  overflow: hidden;
+
+  ${mediaqueries.phablet`
+    -webkit-line-clamp: 3;
+  `}
+`;
+
+const Title = styled(H2)`
+  font-size: 21px;
+  margin-bottom: ${p => (p.hasOverflow && p.isTiles ? '35px' : '10px')};
+  transition: color 0.3s ease-in-out;
+  ${limitToTwoLines};
+
+  ${mediaqueries.desktop`
+    margin-bottom: 15px;
+  `}
+
+  ${mediaqueries.tablet`
+    font-size: 24px;
+  `}
+
+  ${mediaqueries.phablet`
+    font-size: 22px;
+    padding: 30px 20px 0;
+    margin-bottom: 10px;
+    -webkit-line-clamp: 3;
+  `}
+`;
+
+const Excerpt = styled.p<{
+  hasOverflow: boolean;
+  narrow: boolean;
+  isTiles: boolean;
+}>`
+  ${limitToTwoLines};
+  font-size: 16px;
+  margin-bottom: 10px;
+  display: ${p => (p.hasOverflow && p.isTiles ? 'none' : 'box')};
+  max-width: ${p => (p.narrow ? '415px' : '515px')};
+
+  ${mediaqueries.desktop`
+    display: -webkit-box;
+  `}
+
+  ${mediaqueries.phablet`
+    margin-bottom; 15px;
+  `}
+
+  ${mediaqueries.phablet`
+    max-width: 100%;
+    padding:  0 20px;
+    margin-bottom: 20px;
+    -webkit-line-clamp: 3;
+  `}
+`;
+
+const MetaData = styled.div`
+  font-weight: 600;
+  font-size: 16px;
+  opacity: 0.33;
+
+  ${mediaqueries.phablet`
+    max-width: 100%;
+    padding:  0 20px 30px;
+  `}
+`;
