+++ src/sections/article/ArticleNext.tsx
@@ -2,15 +2,12 @@
 import { jsx } from 'theme-ui';
 import React from 'react';
 import styled from '@emotion/styled';
-import { css } from '@emotion/core';
-import { Link } from 'gatsby';
-
-import { H3 } from '@components/Headings';
-import Image from '@components/Image';
 
 import mediaqueries from '@styles/media';
-
 import { IArticle } from '@types';
+import ArticleCard from '@components/ArticleCard';
+import useWindowSize from '@hooks/useWindowWidth';
+import { breakpointMap } from '../../gatsby-plugin-theme-ui';
 
 /**
  * Sits at the bottom of our Article page. Shows the next 2 on desktop and the
@@ -25,101 +22,27 @@ import { IArticle } from '@types';
  * mix into the generic list component.
  */
 const ArticleNext = ({ articles }: { articles: IArticle[] }) => {
-  if (!articles) return null;
+  const windowSize = useWindowSize();
+  if (!articles) {
+    return null;
+  }
+
   const numberOfArticles = articles.length;
   return (
     <Grid numberOfArticles={numberOfArticles}>
-      <GridItem article={articles[0]} />
-      <GridItem article={articles[1]} narrow />
+      <ArticleCard article={articles[0]} isTiles />
+      {windowSize.width >= breakpointMap.tablet && (
+        <ArticleCard article={articles[1]} isTiles narrow />
+      )}
     </Grid>
   );
 };
 
 export default React.memo(ArticleNext);
 
-const GridItem = ({
-  article,
-  narrow,
-}: {
-  article: IArticle;
-  narrow?: boolean;
-}) => {
-  if (!article) return null;
-
-  const hasOverflow = narrow && article.title.length > 35;
-  const imageSource = narrow ? article.hero.narrow : article.hero.regular;
-
-  return (
-    <ArticleLink
-      to={article.slug}
-      data-a11y="false"
-      narrow={narrow ? 'true' : 'false'}
-      sx={{
-        '&:hover h2, &:focus h2': {
-          color: 'accent',
-        },
-        "&[data-a11y='true']:focus::after": {
-          borderColor: 'accent',
-        },
-      }}
-    >
-      <Item
-        sx={{
-          '@media (max-width: 540px)': {
-            backgroundColor: 'card',
-          },
-        }}
-      >
-        <ImageContainer>
-          <Image src={imageSource} />
-        </ImageContainer>
-        <Title
-          dark
-          hasOverflow={hasOverflow}
-          sx={{
-            color: 'primary',
-            fontFamily: 'serif',
-          }}
-        >
-          {article.title}
-        </Title>
-        <Excerpt
-          hasOverflow={hasOverflow}
-          sx={{
-            color: 'grey',
-          }}
-        >
-          {article.excerpt}
-        </Excerpt>
-        <MetaData
-          sx={{
-            color: 'grey',
-          }}
-        >
-          {article.date}
-          {/* · 阅读需要 {article.timeToRead} 分钟 */}
-        </MetaData>
-      </Item>
-    </ArticleLink>
-  );
-};
-
 const wide = '1fr';
 const narrow = '457px';
 
-const limitToTwoLines = css`
-  text-overflow: ellipsis;
-  overflow-wrap: normal;
-  -webkit-line-clamp: 2;
-  -webkit-box-orient: vertical;
-  display: -webkit-box;
-  white-space: normal;
-  overflow: hidden;
-
-  ${mediaqueries.phablet`
-    -webkit-line-clamp: 3;
-  `}
-`;
 const Grid = styled.div<{ numberOfArticles: number }>`
   position: relative;
   display: grid;
@@ -145,136 +68,5 @@ const Grid = styled.div<{ numberOfArticles: number }>`
 
   ${mediaqueries.tablet`
     grid-template-columns: 1fr;
-  `}
-`;
-
-const ImageContainer = styled.div<{ narrow?: boolean }>`
-  position: relative;
-  height: 280px;
-  box-shadow: 0 30px 60px -10px rgba(0, 0, 0, ${p => (p.narrow ? 0.22 : 0.3)}),
-    0 18px 36px -18px rgba(0, 0, 0, ${p => (p.narrow ? 0.25 : 0.33)});
-  margin-bottom: 30px;
-  transition: transform 0.3s var(--ease-out-quad),
-    box-shadow 0.3s var(--ease-out-quad);
-
-  & > div {
-    height: 100%;
-  }
-
-  ${mediaqueries.tablet`
-    height: 220px;
-    margin-bottom: 35px;
-  `}
-
-  ${mediaqueries.phablet`
-    height: 200px;
-    margin-bottom: 0;
-    box-shadow: none;
-    overflow: hidden;
-    border-top-right-radius: 5px;
-    border-top-left-radius: 5px;
-  `}
-`;
-
-const Item = styled.div`
-  position: relative;
-
-  @media (max-width: 540px) {
-    box-shadow: 0px 20px 40px rgba(0, 0, 0, 0.2);
-    border-bottom-right-radius: 5px;
-    border-bottom-left-radius: 5px;
-  }
-`;
-
-const Title = styled(H3)`
-  font-size: 22px;
-  line-height: 1.4;
-  margin-bottom: ${p => (p.hasOverflow ? '45px' : '10px')};
-  transition: color 0.3s ease-in-out;
-  ${limitToTwoLines};
-
-  ${mediaqueries.tablet`
-    margin-bottom: 15px;
-  `}
-  ${mediaqueries.phablet`
-    padding: 30px 20px 0;
-    margin-bottom: 10px;
-    -webkit-line-clamp: 3;
-  `}
-`;
-
-const Excerpt = styled.p<{ narrow?: boolean; hasOverflow?: boolean }>`
-  ${limitToTwoLines};
-  font-size: 16px;
-  margin-bottom: 10px;
-  display: ${p => (p.hasOverflow ? 'none' : 'box')};
-  max-width: ${p => (p.narrow ? '415px' : '515px')};
-
-  ${mediaqueries.desktop`
-    display: -webkit-box;
-  `}
-
-  ${mediaqueries.tablet`
-    margin-bottom: 15px;
-  `}
-
-  ${mediaqueries.phablet`
-    max-width: 100%;
-    padding:  0 20px;
-    margin-bottom: 20px;
-    -webkit-line-clamp: 3;
-  `}
-`;
-
-const MetaData = styled.div`
-  font-weight: 600;
-  font-size: 16px;
-  opacity: 0.33;
-
-  ${mediaqueries.phablet`
-    max-width: 100%;
-    padding:  0 20px 30px;
-  `}
-`;
-
-const ArticleLink = styled(Link)<{ narrow: string }>`
-  position: relative;
-  display: block;
-  width: 100%;
-  height: 100%;
-  top: 0;
-  left: 0;
-  z-index: 1;
-  transition: transform 0.33s var(--ease-out-quart);
-  -webkit-tap-highlight-color: rgba(255, 255, 255, 0);
-
-  &:hover ${ImageContainer} {
-    transform: translateY(-1px);
-    box-shadow: 0 50px 80px -20px rgba(0, 0, 0, 0.27),
-      0 30px 50px -30px rgba(0, 0, 0, 0.3);
-  }
-
-  &[data-a11y='true']:focus::after {
-    content: '';
-    position: absolute;
-    left: -2%;
-    top: -2%;
-    width: 104%;
-    height: 104%;
-    border: 3px solid;
-    background: rgba(255, 255, 255, 0.01);
-  }
-
-  ${p => p.narrow === 'true' && mediaqueries.tablet`display: none;`}
-
-  ${mediaqueries.phablet`
-    &:hover ${ImageContainer} {
-      transform: none;
-      box-shadow: initial;
-    }
-
-    &:active {
-      transform: scale(0.97) translateY(3px);
-    }
   `}
 `;
